+++ src/create-i18n.test.js
@@ -1,11 +1,17 @@
-import createI18n from './create-i18n'
+import createI18n, { getNestedValue } from './create-i18n'
 
 export const testLangData = {
   enUS: {
     hello: 'Hello',
+    nav: {
+      home: 'Home',
+    },
   },
   zhCN: {
     hello: '你好',
+    nav: {
+      home: '首页',
+    },
   },
 }
 
@@ -28,6 +34,7 @@ describe('createI18n', () => {
     i18n.setLang('zhCN')
     expect(i18n.getLang()).toBe('zhCN')
     expect(i18n.t('hello')).toBe('你好')
+    expect(i18n.t('nav.home')).toBe('首页')
 
     const mockOnLanguageChange = jest.fn()
     i18n.listen(mockOnLanguageChange)
@@ -55,5 +62,16 @@ describe('createI18n', () => {
   it('should throw error if call with invalid language data and options', () => {
     expect(() => createI18n(null)).toThrow()
     expect(() => createI18n({}, 123)).toThrow()
+  })
+
+  it('should test different cases of getNestedValue', () => {
+    expect(getNestedValue(testLangData.enUS, ['nav', 'home'])).toBe('Home')
+    expect(getNestedValue(testLangData.zhCN, ['nav', 'home'])).toBe('首页')
+    expect(getNestedValue(testLangData.enUS, ['nav'])).toBe('')
+    expect(getNestedValue(testLangData.enUS, ['hello'])).toBe('Hello')
+    expect(getNestedValue({}, [])).toBe('')
+    expect(getNestedValue({}, [''])).toBe('')
+    expect(getNestedValue({}, ['key'])).toBe('')
+    expect(getNestedValue({ a: 'hi' }, ['key'])).toBe('')
   })
 })