+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0 - Draft
+
+- Support nested language data
+  - `{ enUS: { nav: { home: 'Home' } } }` can be accessed by `t('nav.home')`
+- Support template in `i18n.t` function
+- Support `i18n.addLangData`
+
 ## 1.0.0
 
 - Add argument validation to `createI18n`